Half of my mobile nav refactor work is done. Need to check on scroll positioning and on the inner sub menu rendering.
